requirements.txt
confluent-kafka==2.10.0
pydantic==2.11.5
pydantic-settings==2.9.1
faust-streaming==0.11.3
